Change implementation of static post store from list to sorted-map.
We need to assign IDs to posts to be able to link to them. Lookup of
individual posts will be easier when we store the posts in a map and
use the ID as key.

The ID is also stored inside the value because we need it when
generating links and don't want to pass it around separately.

We use a sorted-map so the posts will be ordered on the main page in a
deterministic way.